requirements.txt
# Core dependencies
requests>=2.28.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
python-dotenv>=0.19.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.6.0
imagehash>=4.3.0

# Machine learning (for future similarity detection)
torch==2.7.1
torchvision==0.22.1
transformers>=4.20.0
clip-by-openai>=1.0

# API integration
google-search-results>=2.4.1  # SerpAPI

# Data handling
pandas>=1.4.0
numpy>=1.21.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.19.0
pytest-cov>=3.0.0

# Development tools
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
